Count carefully at end of buffer in `paredit-count-sexps-forward'.
Ignore-this: 710e264ca330414cfb55c295d0ed5f46

Fixes `C-u M-(' at the top level when there is whitespace at the end
of the buffer: the closing delimiter should go at the end of the last
S-expression, not at the end of the buffer after the whitespace.

Thanks to Eitan Postavsky for the report.

darcs-hash:20110320183854-00fcc-7162d40b9111aed39ca72a913393af8bb2425025