paredit-kill: Tweak rule for adding space if kill-whole-line is t.
For now, disable the yank part of the test -- with kill-whole-line,
killing and then yanking is not always a noop and it's not clear
making it a noop is the right approach.  I don't use kill-whole-line
so I'm not sure what expectations are and I'm reluctant to change it,
except to fix obviously-wrong outcomes like (x|(y)) that should be
(x| (y)).

While here, document the effect of kill-whole-line on paredit-kill.